+++ CHANGELOG.md
@@ -1,1 +1,19 @@
-# TODO
+# Changelog
+
+## [Unreleased]
+
+## [0.1.1] - 2019-03-14
+### Added
+- Documentation
+
+### Fixed
+- standard -> prettier
+- Fix invalid config message
+
+## [0.1.0] - 2019-03-14
+### Added
+- Initial save-hooks
+
+[Unreleased]: https://github.com/oomathias/atom-save-hooks/compare/v0.1.1...master
+[0.1.1]: https://github.com/oomathias/atom-save-hooks/releases/tag/v0.1.1
+[0.1.0]: https://github.com/oomathias/atom-save-hooks/releases/tag/v0.1.0